Add support for menus on windows
Multi window support caused some issues, some of which are yet to be solved. Most visible, opening the command picker on 1 window through a shortcut results in opening the picker on all windows
