+++ src-tauri/src/lib.rs
@@ -4,101 +4,12 @@ mod window;
 
 use crate::scripts::commands::{get_script_commands, reply_editor_request, run_script_command};
 use crate::scripts::loader::scripts::ScriptManager;
-use crate::settings::{get_settings, set_preferred_language, set_theme, set_wrap_lines, Settings};
+use crate::settings::{get_settings, open_settings_window, set_preferred_language, set_theme, set_wrap_lines, Settings};
 use crate::window::{menu, Windows};
 use tauri::async_runtime::{spawn, Mutex};
-use tauri::menu::{AboutMetadataBuilder, Menu, MenuItemBuilder, Submenu, SubmenuBuilder};
 use tauri::path::BaseDirectory;
-use tauri::Error::WebviewLabelAlreadyExists;
-use tauri::{
-    AppHandle, Emitter, Manager, State, WebviewUrl, WebviewWindowBuilder, WindowEvent, Wry,
-};
+use tauri::{Listener, Manager, State, WindowEvent};
 use tauri_plugin_store::StoreExt;
-
-const MENU_ITEM_ID_SETTINGS: &str = "settings";
-const MENU_ITEM_ID_NEW_WINDOW: &str = "window_new";
-const MENU_ITEM_ID_SCRIPTS_OPEN_PICKER: &str = "scripts_open_picker";
-const MENU_ITEM_ID_SCRIPTS_REEXECUTE_LAST: &str = "scripts_reexecute_last";
-
-fn get_file_sub_menu(app: &AppHandle, in_settings: bool) -> Result<Submenu<Wry>, tauri::Error> {
-    let mut builder = SubmenuBuilder::new(app, "File");
-    if !in_settings {
-        let new_window_item = MenuItemBuilder::new("New Window")
-            .id(MENU_ITEM_ID_NEW_WINDOW)
-            .accelerator("CmdOrCtrl+N")
-            .build(app)?;
-        builder = builder.item(&new_window_item).separator();
-    }
-    builder.close_window().build()
-}
-
-fn setup_menu(app: &AppHandle, in_settings: bool) -> Result<(), tauri::Error> {
-    let settings = MenuItemBuilder::new("Settings...")
-        .id(MENU_ITEM_ID_SETTINGS)
-        .accelerator("CmdOrCtrl+,")
-        .enabled(!in_settings)
-        .build(app)?;
-    let app_sub_menu = SubmenuBuilder::new(app, "Snip")
-        .about(Some(
-            AboutMetadataBuilder::new()
-                .name(Some("Snip"))
-                .comments(Some("A simple text editor for quickly editing snippets"))
-                .authors(Some(vec!["Rob Bogie".parse().unwrap()]))
-                .build(),
-        ))
-        .separator()
-        .item(&settings)
-        .separator()
-        .services()
-        .separator()
-        .hide()
-        .hide_others()
-        .show_all()
-        .separator()
-        .quit()
-        .build()?;
-
-    let file_sub_menu = get_file_sub_menu(app, in_settings)?;
-    let app_menu = Menu::new(app)?;
-
-    if !in_settings {
-        let edit_sub_menu = SubmenuBuilder::new(app, "Edit")
-            .undo()
-            .redo()
-            .separator()
-            .cut()
-            .copy()
-            .paste()
-            .select_all()
-            .build()?;
-
-        let open_script_picker_item = MenuItemBuilder::new("Open Picker")
-            .id(MENU_ITEM_ID_SCRIPTS_OPEN_PICKER)
-            .accelerator("CmdOrCtrl+B")
-            .build(app)?;
-        let reexecute_last_script_item = MenuItemBuilder::new("Re-execute Last Script")
-            .id(MENU_ITEM_ID_SCRIPTS_REEXECUTE_LAST)
-            .accelerator("CmdOrCtrl+Shift+B")
-            .enabled(false)
-            .build(app)?;
-        let scripts_sub_menu = SubmenuBuilder::new(app, "Scripts")
-            .item(&open_script_picker_item)
-            .separator()
-            .item(&reexecute_last_script_item)
-            .build()?;
-        app_menu.append_items(&[
-            &app_sub_menu,
-            &file_sub_menu,
-            &edit_sub_menu,
-            &scripts_sub_menu,
-        ])?;
-    } else {
-        app_menu.append_items(&[&app_sub_menu, &file_sub_menu])?;
-    }
-    app_menu.set_as_app_menu()?;
-
-    Ok(())
-}
 
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
@@ -150,10 +61,21 @@ pub fn run() {
             app.manage(Mutex::new(windows));
             app.manage(Mutex::new(script_manager));
 
+            let app_handle = app.handle().clone();
+            app.listen_any("open_settings", move |event| {
+                open_settings_window(&app_handle);
+            });
+            let app_handle = app.handle().clone();
+            app.listen_any("new_window", move |event| {
+                let app_handle = app_handle.clone();
+                spawn(async move {
+                    let windows: State<'_, Mutex<Windows>> = app_handle.state();
+                    let mut windows = windows.lock().await;
+                    windows.create_window(&app_handle).unwrap();
+                });
+            });
+
             menu::initialize_global_handlers(app.handle());
-
-            setup_menu(app.handle(), false)?;
-
             Ok(())
         })
         .on_window_event(|window, event| match event {
@@ -166,8 +88,8 @@ pub fn run() {
                 }
             }
             WindowEvent::Focused(true) => {
-                setup_menu(window.app_handle(), window.label() == "settings")
-                    .expect("Could not replace menu");
+                // setup_menu(window.app_handle(), window.label() == "settings")
+                //     .expect("Could not replace menu");
             }
             _ => {}
         })
