+++ src-tauri/src/window/mod.rs
@@ -72,7 +72,7 @@ impl Windows {
             },
         );
 
-        WebviewWindowBuilder::new(
+        let window = WebviewWindowBuilder::new(
             app_handle,
             id,
             WebviewUrl::App("windows/index.html".parse().unwrap()),
@@ -80,6 +80,8 @@ impl Windows {
         .min_inner_size(800.0, 600.0)
         .title(format!("Snip - Untitled {}", self.last_window_id))
         .build()?;
+
+        menu::on_new_window(&window);
 
         Ok(())
     }
